Added convex mesh loader and covexcollider. The fuckin bug persists.
git-svn-id: https://subversion.assembla.com/svn/ryno@105 b5d08889-fd9f-4148-b882-45fe21c36c0c

+++ trunk/Ryno Engine/Ryno Engine/Headers/Physics/GJK.h
@@ -14,13 +14,13 @@ namespace Ryno{
 		static bool gjk(GameObject* go1, GameObject* go2);
 
 	private:
-		static glm::vec3 support(Collider*, Collider*,glm::vec3& dir);
-		static bool contains_origin( glm::vec3& dir);
-		static bool triangle(glm::vec3& dir);
-		static bool tetrahedron(glm::vec3& dir);
-		static bool check_tetrahedron(const glm::vec3& ao, const glm::vec3& ab, const glm::vec3& ac, const glm::vec3& abc, glm::vec3& dir);
+		static glm::dvec3 support(Collider*, Collider*,glm::dvec3& dir);
+		static bool contains_origin( glm::dvec3& dir);
+		static bool triangle(glm::dvec3& dir);
+		static bool tetrahedron(glm::dvec3& dir);
+		static bool check_tetrahedron(const glm::dvec3& ao, const glm::dvec3& ab, const glm::dvec3& ac, const glm::dvec3& abc, glm::dvec3& dir);
 
-		static glm::vec3 a, b, c, d;	//four simplex points
+		static glm::dvec3 a, b, c, d;	//four simplex points
 		static F32 length; //current number of simplex points
